Fix paragraph retrieval
The problem was that the signature of the function
`highlight_in_paragraph` in `AllData` had changed
but the code in widget was not updated.

This caused an exception in the widget, but the exception
is caught in the widget code. Unfortunately, in the error
handling the message indicating the error was printed between
angled brackets '<' and '>', and upon formatting as HTML
was ignored since it was not a valid HTML tag. Therefore
the error remained invisible. Now the angled brackets were
removed from the error message, and one should be able
to see errors that occur during the paragraph retrieval
in the future.

Some other changes that were done:
- Adjust a logging message text
- Add logging configuration to the widget notebook