Don't load DATA_DIR in load_ee_models_library()
Instead introduce a new parameter `data_and_models_dir` that
replaces the environment variable. Hopefully this makes it
more transparent. The callee of `load_ee_models_library()` now
has to decide where to get the value of this parameter from.

In our codebase we have two places where `load_ee_models_library()`
is called:

1. The mining server. It gets the value from the environment variable
`BBS_DATA_AND_MODELS_DIR`, which can also be defined in the .env file.

2. The create mining cache entrypoint. It has a new parameter
`--data-and-models-dir` which should be used to pass value. If left
blank then the environment variable `BBS_DATA_AND_MODELS_DIR is
checked, and if it's not defined then an error will be thrown.