Set a better entrypoint for bbs_base
The previous entrypoint `bash` was replaced by `env` with
the default argument `bash -l`.

Why `env`? It allows to run any command on the container directly
through docker, e.g. `docker run bbs_base python -V`. This was not
possible with the previous entrypoint. If no arguments are provided
then `docker run bbs_base` falls back to the default argument
`bash -l` and runs an interactive login bash shell.

Why a login (-l) shell? A login shell sources the `~/.profile` file
(a non-login shell doesn't), which, in turn, adds `~/.local/bin` to
the `$PATH` variable. This is necessary for entrypoints of
user-installed python packages to work.