Move from typer to argparse
The checking of the existence of the input paths and the checking of the fact
that the paths do not refer to a directory but regular files instead were
not migrated.

Indeed, argparse does not provide such features as typer does. Implementing
these features ourselves means creating a small library like typer inside
our codebase. This is re-inventing the wheel. That does not make sense.